URI-1177 - Preenchimento de Vetor II

+++ 06-Vetores/62.5._URI-1176/62.5._URI-1176/Program.cs
@@ -12,11 +12,11 @@ namespace _62._5._URI_1176 {
             Int64[] vetFibo = new Int64[T];
 
             for (int i = 0; i < T; i++) {
-                int a = 0;
-                int b = 1;
+                Int64 a = 0;
+                Int64 b = 1;
                 int N = int.Parse(Console.ReadLine());
                 for (int j = 0; j < N; j++) {
-                    int aux = b;
+                    Int64 aux = b;
                     b = a;
                     a = aux + b;
                 }
